Refactor internals to be more accessible / maintainable - Moved non-Express functionality from `WebServer` into `Application` - Moved authentication handling from `WebServer` into `Application` - Moved static-file request handling from `RequestHandler` into `StaticRequestHandler` - Moved view / controller request handling from `RequestHandler` into `DynamicRequestHandler`
