+++ server/api/tests.ts
@@ -3,17 +3,17 @@ import main from "../../server";
 
 export default class Tests {
     public static server: {
-        app: Express.Application;
+        server: Express.Application;
     } = undefined;
 
     public static async init(): Promise<void> {
         if (this.server !== undefined) return;
-        this.server = await main();
+        this.server = (await main()).server;
     }
 
     public static request(): supertest.SuperTest<supertest.Test> {
         if (this.server === undefined) return undefined;
-        return supertest(this.server.app);
+        return supertest(this.server.server);
     }
 
 }
